don't execute query when running to_sql
in AR 4.1 and 4.2, to_sql calls find_with_associations
to attach relevant joins so that the produced SQL matches
the sql actually used when running the query. Bullet's
monkey-patch of find_with_associations calls the original
without a block so it actually passes the query to
the DB connection and runs it.

This commit simply branches the monkey-patch on
block_given? to make it consistent with AR's
definition of the same method. To wit, when
find_with_associations is called by a block, it
yields a new relation with the correct associations
joined and doesn't pass the query on to the DB.