Add response handling for non-Rails Rack responses
When using a setup like e.g. Grape mounted on Rack, the response body
will still be nested under `response.body`. Therefore, just checking for
a `rails?` environment does not work here. Instead, we just check if a
response has a `body` set, and if so, return this. If not, the fallback will
remain to be the `response.first` call.