Make Rails check more specific
In one of our non-Rails projects we added ActionMailer. It comes along
with a number of dependencies and modules namespaced under `Rails`.

The problem is that we don't have a Rails app and therefore `Rails.root`
is not defined. This causes bullet to blow up when getting `Rails.root`.

The fix is to be more specific with the check. I also consolidated the
logic to one method since `rails?` was only ever used to check if Bullet
should call `Rails.root`.