Respect Content-Security-Policy nonces
At the moment, an application with a secure content policy _cannot_
use the alert or console notifiers for Bullet without using a less
secure policy in the development environment, which opens the gates
for developers to forget to secure their _other_ inline scripts, etc.,
and is just a bad practice in general.

When we generate our inline scripts (or `UniformNotifier` adds its
inline scripts to the page), we should apply the script-src nonce from
the response headers, when relevant.

Resolves https://github.com/flyerhzm/bullet/issues/478.