Support Rack::Files::Iterator responses
When using the Rack::Static middleware, the response will be a
Rack::Files::Iterator. Such a response will break
`Bullet::Rack#response_body`, with the result that non-HTML files
will not be served correctly.

Seen when trying to use the Sidekiq admin engine in a Rails
application that also uses Bullet.