Add note about enable resseting detectors
It is not clear to many people that calling Bullet.enable resets all other detectors. This has been an open issue (https://github.com/flyerhzm/bullet/issues/481) for a number of years and doesn't look like being fixed soon. Adding a note in the README would provide a good solution to stop people losing time in the interim (I just lost an hour on this personally).