allow apps to not include the user in a notification
While some may find the current OS user helpful, in other contexts, that user is randomly assigned and not material to the notification itself. Consider a deployment enviroment like Heroku, which will use a different OS user for each deployable, and each deploy.

It may also cause grouping issues in services like Sentry, which use the exception / notification content to determine whether the situation has occurred already.

This does not change the default behavior, but rather introduces a new option to skip including that information in notifications, for those who want to.